Ignore context references in XSLT selection patterns
When analyzing an expression that will be used in a selection pattern, the
presence of things that appear to be context references (unquoted element names,
for example), don’t count as a reference to the context. XSLT selection patterns
don’t use the context in the same way as XPath select expressions.

This fix doesn’t break anything :-) and allows more expressions to be resolved
statically.