Attempt to manage the console output encoding
This PR introduces a new configuration file setting, console-output-encoding.
When XML Calabash writes to the console, it attempts to assure that the output
is appropriate for that encoding. There’s a bit of a messy interaction between
the new “message printer” class and the existing “message reporter” class. More
refactoring may be useful.