Comput a total order over subpipeline steps during graph building
This change replaces the, in retrospect very strange, approach previously taken
where the steps were effectively unordered and at runtime each was tested to see
if it could be run. Now we work out in advance a total order such that no step
runs before any step that it depends on. (With some special cases to deal with
the when/otherwise clauses in a choose.)