Don’t construct a new Saxon configuration for every step
The code was over-zealous when constructing step configurations. It created a
new Saxon Configuration for every step. That’s not necessary; it’s only
necessary that each p:library and p:declare-step (and their descndants) have a
new configuration. Configurations are required at that level because any use of
p:import-functions has to be isolated.

Fix #160