Support -s:stepType to run any known (atomic) step
Fix #236

There are a bunch of changes in here, the builder and DocumentResolverProvider
APIs needed a little refactoring, and the declaration for the cx:unique-id step
turned out to be incomplete. But the major functional change is that you can run
any atomic step directly from the command line with -s:stepType.