Print the underlying cause of failures, even when not in debug mode
Running with --debug show stack traces, which are nice for debugging but
otherwise probably a distraction. With this patch, the underlying cause of
failures is printed even when the stack traces aren’t output. (So, for example,
if a document is not well formed, you’ll get the underlying parser error.)