Fix bugs related to [p:]message
If a message attribute occurs on a declared step, make sure it gets output when
the step is run. (The problem here was namespace confusion at runtime.)

Fix #304

The message attribute is implemented as an option (because that’s the easy way
to work out the value of AVTs with variable and function references), but it
isn’t an option. Don’t accept it in p:with-options.

Fix #308