Experimental implementation of implicit validation of primary inputs
There are a bunch of changes in here.

1. Implicit (XML Schema) validation of primary inputs is controlled by the
validation-mode
2. New configuration properties: validation-mode, try-namespaces,
use-location-hints, cc:xml-schema, and cc:catalog
3. New command line options: --validation-mode, --try-namespaces,
--use-location-hints, --xml-schema, and --catalog
4. New extension attribute, cx:validation-mode that can be applied to steps and
p:declare-step.
5. New elements in p:pipeinfo for loading schemas and catalogs, cx:use-catalog
and cx:import-schema

These are all (perhaps somewhat crudely) documented.

There’s also an update to the trace output. All loaded resources are now
recorded in the trace, along with the time taken to retrieve them.