Support user-styling of the SVG pipeline and graph output
With this PR, you can add a ‘style’ attribute to the `cc:graphviz` configuration
element. That identifies a stylesheet that’s applied to the XML model of the
pipelines and graphs before they’re styled with graphviz. It can do this by
adding dot:* attributes to nodes and edges. Documentation, TBD.