Rework the SVG output for pipelines and graphs
This PR reworks the SVG graph generation more-or-less completely. The XML
description of pipelines and graphs are now broadly similar. Almost all of the
styling is performed by XSLT and there’s an option for users to provide their
own stylesheets. There are some small, pervasive changes that attempt to expose
a little more information in the description.

I added better navigation and on-page scaling to the SVG pages.

There’s a new `--nogo` option that parses a pipeline and builds its graphs
without running it.

This PR mostly resolves #336 but it won’t be very easy to use if there’s no
documentation. That’s next.