Raise an exception of a value template expression has a syntax error
Fix #342

Syntactically invalid AVTs weren’t raising a syntax error and the evaluation
step wasn’t reporting it. This lead to odd dynamic errors. On reflection, none
of the expression errors should be raised at the point of compilation. Instead a
warning is issued and the error only occurs if the expression is evaluated.