Substantial revision to SVG diagrams of pipelines and graphs
Fix #351

This PR restructures the pipeline description markup and how it is styled. This
has no user-visible effects unless you’re interested in styling the SVG
diagrams. That’s substantially changed. The documentation has been updated to
reflect these changes.

The new approach is a little more complicated, but it exposes all of the default
styling to further customization.

There’s an apparent bug in Trang,
https://github.com/relaxng/jing-trang/issues/287 and several of the hacks
required to work around that bug are appalling. I admit it.