Send message output to stderr instead of stdout
Fix #370

This is a user-visible change, but I can’t think of another solution. If you
explicitly request that pipeline output should go to stdout, we can’t also write
messages there. (The fact that we *didn’t* was accidental, but nevermind.)

If you want to capture the pipeline output, you don’t want the messages mixed
in, so stderr is arguably a better destination anyway.