Add content-type support to document caching
Fix #353

This PR updates the way the internal document cache is used so that the same URI
can appear in the cache with different content types. For example, to support
both XML and text inclusions with XInclude. The documentation for the
cx:cache-add and cx:cache-delete steps has been updated to describe how this
works.

This PR also adds the ability to parse media types with wildcards and fixes a
small bug (well, works around a small bug) in pipeline visualization.