Improvements to the p:urify() function
Fix #362

This PR updates the behavior of p:urify(). There are spec changes still under
discussion in this area, but the current behavior is so fundamentally broken
that I’m making a few proactive fixes.

I’ve also arranged it so that every step that asks for a URI value gets the
p:urify()’d result. I’m hoping this improves things on Windows.