Rework the DocumentManager API
Fix #424

This changes some of the configuration options. It’s now possible to provide
your own DocumentManager, ErrorExplainer, or MimetypeFileTypeMap. The latter
replacing the mimetypes map that was previously used.

In addition, you can now set your own EntityResolver, ResourceResolver, or
ModuleURIResolver on the DocumentManager. These are used preferentially. If you
specify your own resolvers, the p:lookup() function will not work. (It can’t
lookup things in the custom resolvers so it fails.)