Improve conformance of p:import-functions
Fix #434

This PR doesn’t use the fix proposed in the issue, but I did refactor the code a
bit so that private functions are excluded and attempts to import multiple
functions with the same name and arity fail. I also sorted out a tricky bug
where the context wasn’t really right inside a declare-step or library that
imported functions. Fixing that bug also allowed me to support nested imports.
Win.