Significant refactoring of message and logging systems
There’s more separation between things that are logged through the
MessageReporter and things that are only logged to the underlying Java logger.
The intent is that things only relevant to debugging XML Calabash go only to the
logger. Things that might be relevant to debugging a pipeline go to both places.

One exception is that details about how the application is configured go only to
the underlying Java logger. This is necessary because we haven’t configured a
MessageReporter before we’ve configured the application!

I’ve simplified the logback configuration and removed the dynamic adjustment of
logging level. Use a different logback.xml configuration file to change the
level. See the chapter on logging for details.

All of the MessageReporter messages do also get logged now. (The documentation
said that was what happened, but it wasn’t.)

I reworked the ErrorExplanation implementation so that it also uses the
MessageReporter, rather than writing directly to stderr.

Off topic: I noticed that the Detail visualizer did not cope well with
threading. I improved it a bit.