Improved error reporting
Fixed a bug where causes sometimes got lost in nested exceptions. When a
pipeline fails, the error report now includes the chain of causes. Separated out
the special case of empty sequence in a type conversion error to improve the
error message. Added a --stacktrace option to enable stack (step) traces in
errors (and disabled them by default). Removed the location information from the
front of each line of explanation when the --explain option is used.